♻️ refactor: CurriculumCard를 편집·조회 컴포넌트로 분리
isEditable 분기가 SortableWorkbookItem과 CurriculumCard 두 곳에 있어
한 파일에 렌더 트리 4개가 공존했고, prop 15개 중 편집 전용 11개가
조회 모드에서는 의미가 없었다.

- CurriculumCardEditable: 드래그 정렬과 입력 편집을 담당
- CurriculumCardReadonly: 펼치기 토글과 읽기 전용 표시를 담당
- CurriculumSummary: 두 모드가 공유하던 워크북·미션 요약을 추출

조회 모드는 useSortable을 disabled 상태로만 쓰고 있어 dnd 의존을 제거했다.
사용처는 생성·수정 페이지가 Editable, 관리 페이지가 Readonly로 나뉜다.
